Update README.md
What is an enduro mountain bike? Should You Buy it?
Should you purchase an enduro bike? The answer hinges on your riding nature and the type of train you have admittance to. If your regular ride is cycling to the top of problematic tracks, then yes. If you like to go wild, crash things as hard as possible, and have a shared contempt for your well-being, the answer is yes. But be truthful with yourself. If you like to ride blue and green trails at a restful speed, you could have more fun on a trail bike. At the end of the day comes down to a personal penchant.